Remove DateTimeConverter and refactor DateConverter
